requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit
tiktoken
faiss-cpu
protobuf~=3.19.0
langchain_google_genai
langchain-community
InstructorEmbedding
huggingface_hub